Improved Contributing.md
Added a few lines about “issues” and doing PRs for absolute beginners. This is meant specially for people looking to make their first contribution, as I think it would help them a lot for easing the workflow and getting started. Please do have a look and do the necessary.
Thanks